extracted OrgEditorController from FullScreenRawEditor
