+++ OrgNoteApp/Flows/OrgRawEditor/FullScreenOrgRawEditorController.storyboard
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Full Screen Org Raw Editor Controller-->
+        <scene sceneID="p5u-96-QuO">
+            <objects>
+                <viewController storyboardIdentifier="FullScreenOrgRawEditorController" id="dA7-LL-83Z" customClass="FullScreenOrgRawEditorController" customModule="OrgNoteApp" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="k5a-ff-L3X">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="tEg-To-nxH">
+                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                            </view>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <constraints>
+                            <constraint firstItem="tEg-To-nxH" firstAttribute="trailing" secondItem="qsa-lj-A0a" secondAttribute="trailing" id="Rf8-HL-C6E"/>
+                            <constraint firstItem="tEg-To-nxH" firstAttribute="bottom" secondItem="qsa-lj-A0a" secondAttribute="bottom" id="WH1-y0-0Ha"/>
+                            <constraint firstItem="tEg-To-nxH" firstAttribute="leading" secondItem="qsa-lj-A0a" secondAttribute="leading" id="c7g-HG-VtF"/>
+                            <constraint firstItem="tEg-To-nxH" firstAttribute="top" secondItem="k5a-ff-L3X" secondAttribute="top" id="eJB-3Q-0VY"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="qsa-lj-A0a"/>
+                    </view>
+                    <connections>
+                        <outlet property="bottomConstraint" destination="WH1-y0-0Ha" id="AMd-Kp-hW0"/>
+                        <outlet property="editorContrinaer" destination="tEg-To-nxH" id="JVh-4E-Btx"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="Mxq-wQ-HYn" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="44" y="118.29085457271366"/>
+        </scene>
+    </scenes>
+</document>
