--- OrgNoteApp/Flows/OrgRawEditor/OrgRawEditorController.storyboard
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Org Raw Editor Controller-->
-        <scene sceneID="p5u-96-QuO">
-            <objects>
-                <viewController storyboardIdentifier="OrgRawEditorController" id="dA7-LL-83Z" customClass="OrgRawEditorController" customModule="OrgNoteApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="k5a-ff-L3X">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" keyboardDismissMode="interactive" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="x1u-s6-VKd">
-                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
-                                <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                                <textInputTraits key="textInputTraits" autocapitalizationType="sentences" spellCheckingType="yes"/>
-                            </textView>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="x1u-s6-VKd" firstAttribute="top" secondItem="k5a-ff-L3X" secondAttribute="top" id="9Bj-tc-CzG"/>
-                            <constraint firstItem="x1u-s6-VKd" firstAttribute="trailing" secondItem="qsa-lj-A0a" secondAttribute="trailing" id="KcM-av-GNK"/>
-                            <constraint firstItem="x1u-s6-VKd" firstAttribute="bottom" secondItem="qsa-lj-A0a" secondAttribute="bottom" id="VbI-yA-02A"/>
-                            <constraint firstItem="x1u-s6-VKd" firstAttribute="leading" secondItem="qsa-lj-A0a" secondAttribute="leading" id="XDg-cQ-rlt"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="qsa-lj-A0a"/>
-                    </view>
-                    <connections>
-                        <outlet property="bottomConstraint" destination="VbI-yA-02A" id="rnH-8Q-XBM"/>
-                        <outlet property="textView" destination="x1u-s6-VKd" id="inA-mQ-UrR"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="Mxq-wQ-HYn" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="44" y="118.29085457271366"/>
-        </scene>
-    </scenes>
-</document>
